Refactor, fix bugs
