+++ src/sagemaker/jumpstart/enums.py
@@ -15,7 +15,7 @@
 from __future__ import absolute_import
 
 from enum import Enum
-from typing import List, Dict, Any
+from typing import List
 
 
 class ModelFramework(str, Enum):
