+++ tests/unit/sagemaker/jumpstart/curated_hub/test_parser_utils.py
@@ -12,7 +12,7 @@
 # language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
-from sagemaker.jumpstart.converter_utils import (
+from sagemaker.jumpstart.curated_hub.parser_utils import (
     camel_to_snake,
     snake_to_upper_camel,
     walk_and_apply_json,
@@ -59,3 +59,20 @@ def test_walk_and_apply_json():
             }
         ],
     }
+
+    assert (
+        walk_and_apply_json(
+            {
+                "notebook_location": "s3://sagemaker-test-objects-do-not-delete/tensorflow-notebooks/tensorflow-ic-bit-s-r101x3-ilsvrc2012-classification-1-inference.ipynb",
+                "dependencies": [
+                    {
+                        "dependency_origin_path": "sagemaker-test-objects-do-not-delete/tensorflow-notebooks/tensorflow-ic-bit-s-r101x3-ilsvrc2012-classification-1-inference.ipynb",
+                        "dependency_copy_path": "sagemaker-hubs-us-west-2-802376408542/default-hub-1667253603.746/Notebook/pentest-3-notebook-1667933000.49/0.0.1/notebook.ipynb",
+                        "dependency_type": "Notebook",
+                    }
+                ],
+            },
+            snake_to_upper_camel,
+        )
+        == BASE_HUB_NOTEBOOK_DOCUMENT
+    )
